+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_validators.py
@@ -30,6 +30,8 @@ def validate_connector_name(namespace):
 
 
 def validate_ssh_key(namespace):
+    if hasattr(namespace, 'no_ssh_key') and namespace.no_ssh_key:
+        return
     string_or_file = (namespace.ssh_key_value or
                       os.path.join(os.path.expanduser('~'), '.ssh', 'id_rsa.pub'))
     content = string_or_file