+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+2.1.0
++++++
+* `az aks create` understands advanced networking (VNet) options
+* `az aks create` accepts options to enable Log Analytics monitoring and HTTP application routing addons
+* `az aks create --no-ssh-key` creates a cluster without using local SSH keys
+* `az aks create --enable-rbac` creates a cluster with Kubernetes Role-Based Access Control
+* `az aks create` handles Azure Active Directory auth options (PREVIEW)
+
 2.0.34
 ++++++
 * `az aks get-credentials` creates the kube config file with more secure filesystem permissions